cambios 100% antes de 2 envio
