--- generate_confusion_matrices.py
@@ -1,386 +1,0 @@
-import tensorflow as tf
-import nump                # Preprocesamiento específico según el modelo
-        if model_type == 'ResNet50V2':
-            # ResNet50V2 usa su propio preprocesamiento
-            image = tf.keras.applications.resnet_v2.preprocess_input(image)
-        else:
-            # Normalización estándar como en el generador original
-            image = image / 255.0
-        
-        # Verificar rango de valores después del preprocesamiento
-        print(f"\n🔍 Estadísticas de la imagen preprocesada ({model_type}):")
-        print(f"  Min: {np.min(image):.4f}")
-        print(f"  Max: {np.max(image):.4f}")
-        print(f"  Mean: {np.mean(image):.4f}")
-        print(f"  Std: {np.std(image):.4f}")
-        
-        return np.expand_dims(image, axis=0)import cv2
-from PIL import Image
-import os
-from sklearn.metrics import confusion_matrix
-import json
-
-# Configuración
-CLASS_NAMES = ['ADI', 'BACK', 'DEB', 'LYM', 'MUC', 'MUS', 'NORM', 'STR', 'TUM']
-
-# Función para preprocesar imagen
-def preprocess_image(image_path, target_size, apply_clahe=False, model_type=None):
-    """Preprocesa una imagen para el modelo usando el mismo preprocesamiento del entrenamiento"""
-    try:
-        # Cargar imagen usando OpenCV como en el generador original
-        image = cv2.imread(image_path)
-        if image is None:
-            print(f"No se pudo cargar la imagen: {image_path}")
-            return None
-        
-        # Convertir BGR a RGB (OpenCV carga en BGR)
-        image = cv2.cvtColor(image, cv2.COLOR_BGR2RGB)
-        
-        # Redimensionar
-        image = cv2.resize(image, target_size)
-        
-        # Aplicar CLAHE si se solicita (útil para modelos híbridos)
-        if apply_clahe:
-            lab = cv2.cvtColor(image, cv2.COLOR_RGB2LAB)
-            l, a, b = cv2.split(lab)
-            clahe = cv2.createCLAHE(clipLimit=2.0, tileGridSize=(8,8))
-            l = clahe.apply(l)
-            lab = cv2.merge([l, a, b])
-            image = cv2.cvtColor(lab, cv2.COLOR_LAB2RGB)
-        
-        # Preprocesamiento específico según el modelo
-        if model_type == 'ResNet50V2':
-            # ResNet50V2 usa su propio preprocesamiento
-            image = tf.keras.applications.resnet_v2.preprocess_input(image)
-        else:
-            # Normalización estándar como en el generador original
-            image = image / 255.0
-        
-        return np.expand_dims(image, axis=0)
-    except Exception as e:
-        print(f"Error procesando imagen {image_path}: {e}")
-        return None
-
-def load_models():
-    """Carga todos los modelos"""
-    print("Cargando modelos...")
-    try:
-        # Configurar logging para suprimir warnings específicos
-        import logging
-        logging.getLogger('absl').setLevel(logging.ERROR)
-        
-        # Función auxiliar para cargar modelos h5
-        def load_h5_model(path):
-            model = tf.keras.models.load_model(path)
-            # Compilar el modelo con métricas básicas para evitar el warning
-            model.compile(optimizer='adam', 
-                        loss='sparse_categorical_crossentropy',
-                        metrics=['accuracy'])
-            return model
-        
-        models = {
-            'CNN Simple': load_h5_model('models/cnn_simple_model.h5'),
-            'ResNet50V2': tf.keras.layers.TFSMLayer(
-                'models/resnet50_model',
-                call_endpoint='serving_default'
-            ),
-            'MobileNetV2 Base': load_h5_model('models/mobilenetv2_base_only.h5'),
-            'Hybrid Attention': load_h5_model('models/Fast_HybridAttention_final.h5'),
-            'Hybrid Autoencoder': load_h5_model('models/Fast_HybridAutoencoder_final.h5')
-        }
-        print("✅ Todos los modelos cargados exitosamente")
-        return models
-    except Exception as e:
-        print(f"❌ Error cargando modelos: {e}")
-        return None
-
-def generate_confusion_matrices_from_validation_data():
-    """Genera matrices de confusión usando datos de validación reales"""
-    
-    # Buscar directorio de validación
-    validation_dirs = [
-        "validation_data/CRC-VAL-HE-20",
-        "../../../CRC-VAL-HE-20",
-        "CRC-VAL-HE-20",
-        "PRUEBAS"  # Como alternativa usar las imágenes de prueba
-    ]
-    
-    validation_dir = None
-    for dir_path in validation_dirs:
-        if os.path.exists(dir_path):
-            validation_dir = dir_path
-            print(f"✅ Usando directorio de validación: {validation_dir}")
-            break
-    
-    if not validation_dir:
-        print("❌ No se encontró directorio de validación")
-        return None, None, None
-    
-    models = load_models()
-    if not models:
-        return None, None, None
-    
-    # Preparar listas para almacenar predicciones y etiquetas reales
-    all_true_labels = []
-    predictions_by_model = {model_name: [] for model_name in models.keys()}
-    
-    print(f"\n📊 Procesando imágenes desde: {validation_dir}")
-    
-    # Si usamos PRUEBAS, procesamos todas las imágenes como diferentes clases
-    if "PRUEBAS" in validation_dir:
-        image_files = [f for f in os.listdir(validation_dir) 
-                        if f.lower().endswith(('.png', '.jpg', '.jpeg'))]
-        
-        for idx, img_file in enumerate(image_files[:50]):  # Limitar a 50 imágenes
-            try:
-                img_path = os.path.join(validation_dir, img_file)
-                # Preprocesar para ResNet50V2
-                img_224_resnet = preprocess_image(img_path, (224, 224), apply_clahe=False, model_type='ResNet50V2')
-                # Preprocesar para otros modelos que usan 224x224
-                img_224 = preprocess_image(img_path, (224, 224), apply_clahe=False, model_type='standard')
-                # Preprocesar para modelos híbridos
-                img_96 = preprocess_image(img_path, (96, 96), apply_clahe=True, model_type='standard')
-                
-                if img_224 is not None and img_96 is not None and img_224_resnet is not None:
-                    # Obtener predicciones de cada modelo primero
-                    current_predictions = {}
-                    all_predictions_valid = True
-                    
-                    for model_name, model in models.items():
-                        try:
-                            if model_name in ['Hybrid Attention', 'Hybrid Autoencoder']:
-                                pred = model.predict(img_96, verbose=0)
-                            elif model_name == 'ResNet50V2':
-                                outputs = model(img_224_resnet, training=False)
-                                pred = tf.cast(outputs[list(outputs.keys())[0]], tf.float32).numpy()
-                                pred = np.expand_dims(pred, axis=0)
-                            else:
-                                pred = model.predict(img_224, verbose=0)
-                            
-                            current_predictions[model_name] = np.argmax(pred[0])
-                        except Exception as e:
-                            print(f"❌ Error con modelo {model_name} en imagen {img_file}: {e}")
-                            all_predictions_valid = False
-                            break
-                    
-                    # Solo agregar la etiqueta real si todas las predicciones fueron exitosas
-                    if all_predictions_valid:
-                        class_idx = idx % len(CLASS_NAMES)
-                        all_true_labels.append(class_idx)
-                        for model_name, prediction in current_predictions.items():
-                            predictions_by_model[model_name].append(prediction)
-                    
-                    print(f"✅ Procesada imagen {idx+1}/{len(image_files[:50])}: {img_file}")
-                    
-            except Exception as e:
-                print(f"❌ Error procesando {img_file}: {e}")
-                continue
-    else:
-        # Procesar por clases organizadas en carpetas
-        for class_idx, class_name in enumerate(CLASS_NAMES):
-            class_dir = os.path.join(validation_dir, class_name)
-            if not os.path.exists(class_dir):
-                print(f"⚠️ No se encuentra directorio para clase {class_name}")
-                continue
-            
-            image_files = [f for f in os.listdir(class_dir) 
-                          if f.lower().endswith(('.png', '.jpg', '.jpeg', '.tif', '.tiff'))]
-            
-            # Maximo 50 imágenes por clase
-            image_files = image_files[:10]
-            print(f"📂 Procesando clase {class_name}: {len(image_files)} imágenes")
-            
-            for img_file in image_files:
-                try:
-                    img_path = os.path.join(class_dir, img_file)
-                    # Preprocesar para ResNet50V2
-                    img_224_resnet = preprocess_image(img_path, (224, 224), apply_clahe=False, model_type='ResNet50V2')
-                    # Preprocesar para otros modelos que usan 224x224
-                    img_224 = preprocess_image(img_path, (224, 224), apply_clahe=False, model_type='standard')
-                    # Preprocesar para modelos híbridos
-                    img_96 = preprocess_image(img_path, (96, 96), apply_clahe=True, model_type='standard')
-                    
-                    if img_224 is not None and img_96 is not None and img_224_resnet is not None:
-                        # Obtener predicciones de cada modelo primero
-                        current_predictions = {}
-                        all_predictions_valid = True
-                        
-                        for model_name, model in models.items():
-                            try:
-                                if model_name in ['Hybrid Attention', 'Hybrid Autoencoder']:
-                                    pred = model.predict(img_96, verbose=0)
-                                elif model_name == 'ResNet50V2':
-                                    outputs = model(img_224_resnet, training=False)
-                                    pred = outputs[list(outputs.keys())[0]].numpy()
-                                    pred = np.expand_dims(pred, axis=0)
-                                else:
-                                    pred = model.predict(img_224, verbose=0)
-                                
-                                # Aplicar softmax para obtener probabilidades
-                                probabilities = tf.nn.softmax(pred[0]).numpy()
-                                predicted_class = np.argmax(probabilities)
-                                
-                                # Imprimir diagnóstico detallado para cada predicción
-                                print(f"\n📊 Predicción para {model_name} en {img_file}:")
-                                print(f"Clase predicha: {predicted_class} ({CLASS_NAMES[predicted_class]})")
-                                print(f"Probabilidad: {probabilities[predicted_class]:.4f}")
-                                print(f"Top 3 probabilidades:")
-                                top_k = np.argsort(probabilities)[-3:][::-1]
-                                for idx in top_k:
-                                    print(f"  {CLASS_NAMES[idx]}: {probabilities[idx]:.4f}")
-                                
-                                current_predictions[model_name] = predicted_class
-                            except Exception as e:
-                                print(f"❌ Error con modelo {model_name}: {e}")
-                                all_predictions_valid = False
-                                break
-                        
-                        # Solo agregar la etiqueta real si todas las predicciones fueron exitosas
-                        if all_predictions_valid:
-                            all_true_labels.append(class_idx)
-                            for model_name, prediction in current_predictions.items():
-                                predictions_by_model[model_name].append(prediction)
-                                try:    
-                                    # Mostrar diagnóstico si hay un error en la predicción
-                                    predicted_class = np.argmax(pred[0])
-                                    if predicted_class != class_idx:
-                                        probabilities = tf.nn.softmax(pred[0]).numpy()
-                                        print(f"\n⚠️ Predicción incorrecta en {model_name} para {img_file}:")
-                                        print(f"Clase real: {class_idx} ({CLASS_NAMES[class_idx]})")
-                                        print(f"Clase predicha: {predicted_class} ({CLASS_NAMES[predicted_class]})")
-                                        print(f"Probabilidad: {probabilities[predicted_class]:.4f}")
-                                        
-                                except Exception as e:
-                                    print(f"❌ Error con modelo {model_name}: {e}")
-                                    predictions_by_model[model_name].append(class_idx)  # Usar etiqueta real como fallback
-                        
-                except Exception as e:
-                    print(f"❌ Error procesando {img_file}: {e}")
-                    continue
-    
-    # Calcular matrices de confusión
-    confusion_matrices = {}
-    accuracies = {}
-    
-    print(f"\n🔄 Calculando matrices de confusión...")
-    print(f"Total de muestras procesadas: {len(all_true_labels)}")
-    
-    for model_name in models.keys():
-        if len(predictions_by_model[model_name]) > 0:
-            conf_matrix = confusion_matrix(
-                all_true_labels,
-                predictions_by_model[model_name],
-                labels=range(len(CLASS_NAMES))
-            )
-            confusion_matrices[model_name] = conf_matrix
-            
-            # Calcular accuracy
-            accuracy = np.sum(np.diag(conf_matrix)) / np.sum(conf_matrix)
-            accuracies[model_name] = accuracy
-            
-            print(f"✅ {model_name}: Accuracy = {accuracy*100:.2f}%")
-        else:
-            print(f"❌ No hay predicciones para {model_name}")
-    
-    return confusion_matrices, accuracies, models
-
-def generate_python_code_for_matrices(confusion_matrices, accuracies):
-    """Genera el código Python con las matrices de confusión reales"""
-    
-    print("\n" + "="*80)
-    print("📋 CÓDIGO PARA REEMPLAZAR EN app.py:")
-    print("="*80)
-    
-    # Generar código para matrices de confusión
-    code = "        # Matrices de confusión calculadas con datos reales\n"
-    code += "        confusion_matrices = {\n"
-    
-    for model_name, matrix in confusion_matrices.items():
-        code += f"            '{model_name}': np.array(["
-        
-        for i, row in enumerate(matrix):
-            if i == 0:
-                code += "["
-            else:
-                code += "                                  ["
-            
-            code += ", ".join([f"{val:3d}" for val in row])
-            
-            if i == len(matrix) - 1:
-                code += "]]),\n"
-            else:
-                code += "],\n"
-    
-    code += "        }\n\n"
-    
-    # Generar código para accuracies
-    code += "        # Accuracies reales calculadas\n"
-    code += "        accuracies = {\n"
-    for model_name, acc in accuracies.items():
-        code += f"            '{model_name}': {acc:.4f},  # {acc*100:.2f}%\n"
-    code += "        }\n\n"
-    
-    # Generar código para losses (estimadas basadas en accuracy)
-    code += "        # Losses estimadas (basadas en accuracy)\n"
-    code += "        losses = {\n"
-    for model_name, acc in accuracies.items():
-        # Estimar loss basada en accuracy (relación inversa aproximada)
-        estimated_loss = max(0.1, (1 - acc) * 2)
-        code += f"            '{model_name}': {estimated_loss:.4f},\n"
-    code += "        }"
-    
-    print(code)
-    print("="*80)
-    
-    # Guardar en archivo
-    with open("confusion_matrices_code.txt", "w", encoding='utf-8') as f:
-        f.write(code)
-    
-    print("✅ Código guardado en 'confusion_matrices_code.txt'")
-    
-    return code
-
-def save_matrices_as_json(confusion_matrices, accuracies):
-    """Guarda las matrices en formato JSON para referencia"""
-    
-    # Convertir matrices numpy a listas para JSON
-    matrices_dict = {}
-    for model_name, matrix in confusion_matrices.items():
-        matrices_dict[model_name] = {
-            'confusion_matrix': matrix.tolist(),
-            'accuracy': float(accuracies[model_name])
-        }
-    
-    with open("confusion_matrices_data.json", "w", encoding='utf-8') as f:
-        json.dump(matrices_dict, f, indent=2, ensure_ascii=False)
-    
-    print("✅ Matrices guardadas en 'confusion_matrices_data.json'")
-
-def main():
-    print("🚀 Generando matrices de confusión reales...")
-    print("="*60)
-    
-    confusion_matrices, accuracies, models = generate_confusion_matrices_from_validation_data()
-    
-    if confusion_matrices:
-        print(f"\n📊 Resumen de resultados:")
-        print("-" * 40)
-        for model_name, accuracy in accuracies.items():
-            print(f"{model_name:20}: {accuracy*100:6.2f}%")
-        
-        # Generar código Python
-        generate_python_code_for_matrices(confusion_matrices, accuracies)
-        
-        # Guardar en JSON
-        save_matrices_as_json(confusion_matrices, accuracies)
-        
-        print(f"\n✅ ¡Proceso completado exitosamente!")
-        print(f"📁 Archivos generados:")
-        print(f"   - confusion_matrices_code.txt (código para app.py)")
-        print(f"   - confusion_matrices_data.json (datos en JSON)")
-        
-    else:
-        print("❌ No se pudieron generar las matrices de confusión")
-
-if __name__ == "__main__":
-    main()
