+++ reports/validation_summary.csv
@@ -1,0 +1,7 @@
+Modelo,Accuracy,Top-3 Acc,Precision,Recall,F1-Score,Tiempo (s),Muestras
+CNN_Simple,0.4994 (49.94%),0.7198 (71.98%),0.4839,0.4994,0.4554,90.40,7180
+ResNet50,0.5712 (57.12%),0.8453 (84.53%),0.5356,0.5712,0.5434,751.64,7180
+MobileNetV2,0.9145 (91.45%),0.9925 (99.25%),0.9189,0.9145,0.9161,227.34,7180
+HybridAttention,0.1717 (17.17%),0.3425 (34.25%),0.0295,0.1717,0.0503,70.57,7180
+HybridAutoencoder,0.1717 (17.17%),0.3014 (30.14%),0.0295,0.1717,0.0503,97.82,7180
+ResNet50 (con CLAHE),0.5174 (51.74%),0.8209 (82.09%),0.4932,0.5174,0.4771,630.03,7180
